refactor: FSD 레이어 위반 정리 (shared→features 제거, Header widgets 편입)
- 실시간 동기화(BookmarkRealtimeSync/ConditionalRealtimeSync)를 shared/lib
  → features/bookmark/ui 로 이동. shared가 features를 import하던 유일한
  역방향 위반 제거 (shared의 "의존성 없는 레이어" 전제 회복).
- Header를 FSD 밖 components/layout → widgets/layout 으로 편입, 빈 components/ 삭제.
- import 갱신은 app/(main)/layout.tsx 한 곳뿐.
